Move should_navigation_request_be_blocked out of GlobalScope
